Refactor: rename AuthSection folder to auth
