+++ src/app/auth-v2/page.tsx
@@ -1,4 +1,4 @@
-import AuthSectionV2 from "@/sections/AuthSection/AuthSectionV2";
+import AuthSectionV2 from "@/sections/auth/AuthSectionV2";
 
 export default function AuthPage() {
   return <AuthSectionV2 />;
