+++ src/sections/auth/elements/AuthTextV2.tsx
